requirements.txt
python-dotenv
pandas
numpy
matplotlib
pymongo
certifi 
scikit-learn
seaborn
pymongo[srv]
scikit-learn
dill
pyaml

#-e .
